examples : refactor common code into a library
